Add files via upload
I analyzed a data set with QS University rankings from 2017 to 2022. I uncovered trends in the data, like Countries with the highest University to Population Ratio, Average University Ratings by Country, Cities, Countries with the Highest International Students, Research Output by Region, and the trends of ratings for different regions over the year. I used the libraries of Pandas, BeautifulSoup, Maplotlib, and Seaborn to clean and analyze the dataset. I also used excel to remove the empty rows and columns of the dataset and drop the unecessary rows in the dataset. The original dataset alongwith the Pictures of the Power Bi dashboards and the Jupyter notebook for the project are shown here.